% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_bias.R
\name{boundary_bias}
\alias{boundary_bias}
\title{Bias on the boundary}
\usage{
boundary_bias(x0 = 0, dat, kernID = 0, left = TRUE, maxit = 20, tol =
  1e-3)
}
\arguments{
\item{x0}{A scalar, boundary point in estimation.}

\item{dat}{A list with the following components: \describe{ \item{y}{A vector
treatment outcomes. In a fuzzy RD, this variable can also be a vector of
treatment assignment variables.} \item{x}{A vector of covariates.}
\item{h}{A scalar bandwidth.} \item{p}{The polynomial degree. p = 2 is used
to estimate the second derivative of the conditional mean function for bias
estimation. LCQR is used to estimate the second derivative.}
\item{q}{Number of quantiles used in estimation. It needs to be an odd
integer such as 5 or 9.} }}

\item{kernID}{Kernel id number. \enumerate{ \item \code{kernID = 0}:
triangular kernel. \item \code{kernID = 1}: biweight kernel. \item
\code{kernID = 2}: Epanechnikov kernel. \item \code{kernID = 3}: Gaussian
kernel. \item \code{kernID = 4}: tricube kernel. \item \code{kernID = 5}:
triweight kernel. \item \code{kernID = 6}: uniform kernel. }}

\item{left}{A logical variable that takes the value \code{TRUE} for data to
the left of (below) the cutoff. Defaults to \code{TRUE}.}

\item{maxit}{Maximum iteration number in the MM algorithm for quantile
estimation. Defaults to 20.}

\item{tol}{Convergence criterion in the MM algorithm. Defaults to 1.0e-3.}
}
\value{
Estimated bias on the boundary.
}
\description{
This function computes the bias on the boundary, typicall at x =
0, in a local composite quantile regression.
}
