% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headstart}
\alias{headstart}
\title{The Head Start dataset from Ludwig and Miller (2007)}
\format{
A data frame with 3,103 rows and 2 variables: \describe{
\item{\code{mortality}}{the treatment outcome variable \code{y}, morality
rates per 10,000 children between 5 and 9 years old between 1973-1983.}
\item{\code{poverty}}{the covariate \code{x}, a county's poverty rate in
1960 relative to the 300th poorest county (59.198)} }
}
\usage{
headstart
}
\description{
The Head Start policy dataset on the mortality rates for
children
}
\details{
Ludwig and Miller (2007) studies the impact of Head Start funding on
children's schooling and health. This subset of their data focuses on the
change in mortality rates for children due to the Head Start funding cutoff
set by the Office of Economic Opportunities (OEF). Changes in morality
rates near the funding cutoff can be used to identify the causal effect of
such policy. We subtract the average poverty rate (59.198) from the
poverty rate data so that the cutoff is 0 for treatment assignment. Each
row in the data frame represents a county.
}
\references{
{

\cite{Ludwig, J. and D. L. MIller (2007) "Does Head Start improve children's
life chances? Evidence from a regression discontinuity design," Quarterly
Journal of Economics, 122 (1), 159-208.}

}
}
\keyword{datasets}
