% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee}
\alias{lee}
\title{US House elections dataset from Lee (2008)}
\format{
A data frame with 6,558 rows and 2 variables. \describe{
\item{\code{voteshare}}{the treatment outcome variable \code{y}, vote
share in the next election} \item{\code{margin}}{the covariate \code{x},
democratic margin of victory at the previous election} } Each row
}
\source{
Mostly Harmless Econometrics data archive at
\url{https://economics.mit.edu/faculty/angrist/data1/mhe}
}
\usage{
lee
}
\description{
Data on elections to the U.S. House of Representatives
}
\details{
Lee (2008) uses the data on elections to the U.S. House of
Representatives  (1946-1998) to analyze the incumbency advantage. For a
party's candidate who barely won the election (and became incumbent) or
barely lost the previous election, their districts usually share many
common characteristics. The electoral success for the two groups of common
characteristics. The electoral success for the two groups of candidates in
the next election can be used to identify the causal incumbency advantage.
We subtract the number 50, equivalent to fifty percent, from the vote share
data so that the cutoff is 0 for treatment assignment. Each row of the data
frame represents a party's candidate.
}
\references{
{

\cite{Lee, D. S. (2008) "Randomized experiments from non-random selection in
U.S. House elections," Journal of Econometrics, 142 (2), 675-697.}

}
}
\keyword{datasets}
