\name{rdcqr-package}
\alias{rdcqr-package}
\docType{package}
\title{
\packageTitle{rdcqr}
}
\description{
\packageDescription{rdcqr}
}

\author{
\packageAuthor{rdcqr}

Maintainer: \packageMaintainer{rdcqr}
}

\keyword{ rdcqr }
\seealso{
The package and its update can be found at 
\url{https://github.com/xhuang20/rdcqr}.
}

