\name{rdcqr-package}
\alias{rdcqr-package}
\docType{package}
\title{
\packageTitle{rdcqr}
}
\description{
\packageDescription{rdcqr}
}

\author{
\packageAuthor{rdcqr}

Maintainer: \packageMaintainer{rdcqr}
}

\keyword{ rdcqr }
\seealso{
Optional links to other man pages, e.g.
\url{http://www.r-project.org}.
}

